Final Revision Code (Not Tested On Robot)
