+++ Src/main.c
@@ -116,6 +116,7 @@ int main(void)
   MX_TIM13_Init();
   MX_TIM14_Init();
   /* USER CODE BEGIN 2 */
+  Putty_Init();
 
   /* USER CODE END 2 */
 
@@ -128,7 +129,7 @@ int main(void)
   /* USER CODE END WHILE */
 
   /* USER CODE BEGIN 3 */
-
+  Putty_Callback();
   }
   /* USER CODE END 3 */
 
@@ -193,7 +194,11 @@ void SystemClock_Config(void)
 }
 
 /* USER CODE BEGIN 4 */
-
+void HAL_UART_RxCpltCallback(UART_HandleTypeDef *huart){
+	if(huart->Instance == huart3.Instance){//input from the PC
+		puttystruct.huart_Rx_len = 1;
+		puttystruct.small_buf[0] = *(huart->pRxBuffPtr-1);
+}
 /* USER CODE END 4 */
 
 /**
